+++ src/ServiceStack.Interfaces/Redis/IRedisNativeClient.cs
@@ -20,8 +20,8 @@ namespace ServiceStack.Redis
 	{
 		//Redis utility operations
 		Dictionary<string, string> Info { get; }
-		int Db { get; set; }
-		int DbSize { get; }
+		long Db { get; set; }
+		long DbSize { get; }
 		DateTime LastSave { get; }
 		void Save();
 		void BgSave();
@@ -43,17 +43,17 @@ namespace ServiceStack.Redis
 		byte[] Restore(string key, long expireMs, byte[] dumpValue);
 		void Migrate(string host, int port, int destinationDb, long timeoutMs);
 		bool Move(string key, int db);
-		int ObjectIdleTime(string key);
+		long ObjectIdleTime(string key);
 
 		//Common key-value Redis operations
 		byte[][] Keys(string pattern);
-		int Exists(string key);
-		int StrLen(string key);
+		long Exists(string key);
+		long StrLen(string key);
 		void Set(string key, byte[] value);
 		void SetEx(string key, int expireInSeconds, byte[] value);
 		bool Persist(string key);
 		void PSetEx(string key, long expireInMs, byte[] value);
-		int SetNX(string key, byte[] value);
+		long SetNX(string key, byte[] value);
 		void MSet(byte[][] keys, byte[][] values);
 		void MSet(string[] keys, byte[][] values);
 		bool MSetNx(byte[][] keys, byte[][] values);
@@ -62,20 +62,20 @@ namespace ServiceStack.Redis
 		byte[] GetSet(string key, byte[] value);
 		byte[][] MGet(params byte[][] keysAndArgs);
 		byte[][] MGet(params string[] keys);
-		int Del(string key);
-		int Del(params string[] keys);
+		long Del(string key);
+		long Del(params string[] keys);
 		long Incr(string key);
 		long IncrBy(string key, int incrBy);
 		double IncrByFloat(string key, double incrBy);
 		long Decr(string key);
 		long DecrBy(string key, int decrBy);
-		int Append(string key, byte[] value);
+		long Append(string key, byte[] value);
 		[Obsolete("Was renamed to GetRange in 2.4")]
 		byte[] Substr(string key, int fromIndex, int toIndex);
 		byte[] GetRange(string key, int fromIndex, int toIndex);
-		int SetRange(string key, int offset, byte[] value);
-		int GetBit(string key, int offset);
-		int SetBit(string key, int offset, int value);
+		long SetRange(string key, int offset, byte[] value);
+		long GetBit(string key, int offset);
+		long SetBit(string key, int offset, int value);
 		
 		string RandomKey();
 		void Rename(string oldKeyname, string newKeyname);
@@ -84,7 +84,7 @@ namespace ServiceStack.Redis
 		bool PExpire(string key, long ttlMs);
 		bool ExpireAt(string key, long unixTime);
 		bool PExpireAt(string key, long unixTimeMs);
-		int Ttl(string key);
+		long Ttl(string key);
 		long PTtl(string key);
 
 		//Redis Sort operation (works on lists, sets or hashes)
@@ -92,13 +92,13 @@ namespace ServiceStack.Redis
 
 		//Redis List operations
 		byte[][] LRange(string listId, int startingFrom, int endingAt);
-		int RPush(string listId, byte[] value);
-		int RPushX(string listId, byte[] value);
-		int LPush(string listId, byte[] value);
-		int LPushX(string listId, byte[] value);
+		long RPush(string listId, byte[] value);
+		long RPushX(string listId, byte[] value);
+		long LPush(string listId, byte[] value);
+		long LPushX(string listId, byte[] value);
 		void LTrim(string listId, int keepStartingFrom, int keepEndingAt);
-		int LRem(string listId, int removeNoOfMatches, byte[] value);
-		int LLen(string listId);
+		long LRem(string listId, int removeNoOfMatches, byte[] value);
+		long LLen(string listId);
 		byte[] LIndex(string listId, int listIndex);
         void LInsert(string listId, bool insertBefore, byte[] pivot, byte[] value);
 		void LSet(string listId, int listIndex, byte[] value);
@@ -116,12 +116,12 @@ namespace ServiceStack.Redis
 
 		//Redis Set operations
 		byte[][] SMembers(string setId);
-		int SAdd(string setId, byte[] value);
-		int SRem(string setId, byte[] value);
+		long SAdd(string setId, byte[] value);
+		long SRem(string setId, byte[] value);
 		byte[] SPop(string setId);
 		void SMove(string fromSetId, string toSetId, byte[] value);
-		int SCard(string setId);
-		int SIsMember(string setId, byte[] value);
+		long SCard(string setId);
+		long SIsMember(string setId, byte[] value);
 		byte[][] SInter(params string[] setIds);
 		void SInterStore(string intoSetId, params string[] setIds);
 		byte[][] SUnion(params string[] setIds);
@@ -132,13 +132,13 @@ namespace ServiceStack.Redis
 
 
 		//Redis Sorted Set operations
-		int ZAdd(string setId, double score, byte[] value);
-		int ZAdd(string setId, long score, byte[] value);
-		int ZRem(string setId, byte[] value);
+		long ZAdd(string setId, double score, byte[] value);
+		long ZAdd(string setId, long score, byte[] value);
+		long ZRem(string setId, byte[] value);
 		double ZIncrBy(string setId, double incrBy, byte[] value);
 		double ZIncrBy(string setId, long incrBy, byte[] value);
-		int ZRank(string setId, byte[] value);
-		int ZRevRank(string setId, byte[] value);
+		long ZRank(string setId, byte[] value);
+		long ZRevRank(string setId, byte[] value);
 		byte[][] ZRange(string setId, int min, int max);
 		byte[][] ZRangeWithScores(string setId, int min, int max);
 		byte[][] ZRevRange(string setId, int min, int max);
@@ -151,25 +151,25 @@ namespace ServiceStack.Redis
 		byte[][] ZRevRangeByScore(string setId, long min, long max, int? skip, int? take);
 		byte[][] ZRevRangeByScoreWithScores(string setId, double min, double max, int? skip, int? take);
 		byte[][] ZRevRangeByScoreWithScores(string setId, long min, long max, int? skip, int? take);
-		int ZRemRangeByRank(string setId, int min, int max);
-		int ZRemRangeByScore(string setId, double fromScore, double toScore);
-		int ZRemRangeByScore(string setId, long fromScore, long toScore);
-		int ZCard(string setId);
+		long ZRemRangeByRank(string setId, int min, int max);
+		long ZRemRangeByScore(string setId, double fromScore, double toScore);
+		long ZRemRangeByScore(string setId, long fromScore, long toScore);
+		long ZCard(string setId);
 		double ZScore(string setId, byte[] value);
-		int ZUnionStore(string intoSetId, params string[] setIds);
-		int ZInterStore(string intoSetId, params string[] setIds);
+		long ZUnionStore(string intoSetId, params string[] setIds);
+		long ZInterStore(string intoSetId, params string[] setIds);
 
 		//Redis Hash operations
-		int HSet(string hashId, byte[] key, byte[] value);
+		long HSet(string hashId, byte[] key, byte[] value);
 		void HMSet(string hashId, byte[][] keys, byte[][] values);
-		int HSetNX(string hashId, byte[] key, byte[] value);
-		int HIncrby(string hashId, byte[] key, int incrementBy);
+		long HSetNX(string hashId, byte[] key, byte[] value);
+		long HIncrby(string hashId, byte[] key, int incrementBy);
 		double HIncrbyFloat(string hashId, byte[] key, double incrementBy);
 		byte[] HGet(string hashId, byte[] key);
 		byte[][] HMGet(string hashId, params byte[][] keysAndArgs);
-		int HDel(string hashId, byte[] key);
-		int HExists(string hashId, byte[] key);
-		int HLen(string hashId);
+		long HDel(string hashId, byte[] key);
+		long HExists(string hashId, byte[] key);
+		long HLen(string hashId);
 		byte[][] HKeys(string hashId);
 		byte[][] HVals(string hashId);
 		byte[][] HGetAll(string hashId);
@@ -177,7 +177,7 @@ namespace ServiceStack.Redis
 		//Redis Pub/Sub operations
 		void Watch(params string[] keys);
 		void UnWatch();
-		int Publish(string toChannel, byte[] message);
+		long Publish(string toChannel, byte[] message);
 		byte[][] Subscribe(params string[] toChannels);
 		byte[][] UnSubscribe(params string[] toChannels);
 		byte[][] PSubscribe(params string[] toChannelsMatchingPatterns);
@@ -185,8 +185,8 @@ namespace ServiceStack.Redis
 		byte[][] ReceiveMessages();
 
         //Redis LUA support
-        int EvalInt(string luaBody, int numberOfKeys, params byte[][] keysAndArgs);
-        int EvalShaInt(string sha1, int numberOfKeys, params byte[][] keysAndArgs);
+		long EvalInt(string luaBody, int numberOfKeys, params byte[][] keysAndArgs);
+		long EvalShaInt(string sha1, int numberOfKeys, params byte[][] keysAndArgs);
         string EvalStr(string luaBody, int numberOfKeys, params byte[][] keysAndArgs);
         string EvalShaStr(string sha1, int numberOfKeys, params byte[][] keysAndArgs);
         byte[][] Eval(string luaBody, int numberOfKeys, params byte[][] keysAndArgs);
